+++ packages/core-mobile/app/screens/bridge/hooks/useBridgeNetworks.ts
@@ -1,6 +1,7 @@
-import { useNetworksFromCaip2ChainIds } from 'temp/caip2ChainIds'
 import { Network } from '@avalabs/core-chains-sdk'
 import { BridgeAsset } from '@avalabs/bridge-unified'
+import { useNetworks } from 'hooks/networks/useNetworks'
+import { getChainIdFromCaip2 } from 'temp/caip2ChainIds'
 import { useUnifiedBridgeAssets } from './useUnifiedBridgeAssets'
 
 export const useBridgeSourceNetworks = (): Network[] => {
@@ -15,4 +16,16 @@ export const useBridgeTargetNetworks = (
   return useNetworksFromCaip2ChainIds(
     Object.keys(selectedBridgeAsset?.destinations ?? [])
   )
+}
+
+export const useNetworksFromCaip2ChainIds = (
+  caip2ChainIds: string[]
+): Network[] => {
+  const { networks } = useNetworks()
+
+  return caip2ChainIds
+    .map(caip2ChainId => getChainIdFromCaip2(caip2ChainId))
+    .filter((chainId): chainId is number => !!chainId)
+    .map(chainId => networks[chainId])
+    .filter((network): network is Network => !!network)
 }
